requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
click==8.1.7
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.1.4
python-dateutil==2.8.2
typing-extensions==4.8.0
